delete if exists

+++ packages/backend/scripts/tvl-restore/clear-tvl.sql
@@ -5,6 +5,8 @@ DECLARE
 BEGIN
     FOREACH table_name IN ARRAY tables
     LOOP
-        EXECUTE format('DELETE FROM %I', table_name);
+        IF EXISTS (SELECT FROM pg_tables WHERE schemaname = 'public' AND tablename = table_name) THEN
+            EXECUTE format('DELETE FROM %I', table_name);
+        END IF;
     END LOOP;
 END $$;
